more comments on collections
